# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Task Manager SaaS API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Project {
  id: number;
  userId: string;
  name: string;
  /** @nullable */
  description?: string | null;
  color: string;
  taskCount: number;
  completedCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateProjectBody {
  name: string;
  /** @nullable */
  description?: string | null;
  color: string;
}

export interface UpdateProjectBody {
  name?: string;
  /** @nullable */
  description?: string | null;
  color?: string;
}

export type TaskStatus = (typeof TaskStatus)[keyof typeof TaskStatus];

export const TaskStatus = {
  todo: "todo",
  in_progress: "in_progress",
  done: "done",
} as const;

export type TaskPriority = (typeof TaskPriority)[keyof typeof TaskPriority];

export const TaskPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface Task {
  id: number;
  userId: string;
  /** @nullable */
  projectId?: number | null;
  /** @nullable */
  projectName?: string | null;
  /** @nullable */
  projectColor?: string | null;
  title: string;
  /** @nullable */
  description?: string | null;
  status: TaskStatus;
  priority: TaskPriority;
  /** @nullable */
  dueDate?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type CreateTaskBodyStatus =
  (typeof CreateTaskBodyStatus)[keyof typeof CreateTaskBodyStatus];

export const CreateTaskBodyStatus = {
  todo: "todo",
  in_progress: "in_progress",
  done: "done",
} as const;

export type CreateTaskBodyPriority =
  (typeof CreateTaskBodyPriority)[keyof typeof CreateTaskBodyPriority];

export const CreateTaskBodyPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface CreateTaskBody {
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  projectId?: number | null;
  status?: CreateTaskBodyStatus;
  priority?: CreateTaskBodyPriority;
  /** @nullable */
  dueDate?: string | null;
}

export type UpdateTaskBodyStatus =
  (typeof UpdateTaskBodyStatus)[keyof typeof UpdateTaskBodyStatus];

export const UpdateTaskBodyStatus = {
  todo: "todo",
  in_progress: "in_progress",
  done: "done",
} as const;

export type UpdateTaskBodyPriority =
  (typeof UpdateTaskBodyPriority)[keyof typeof UpdateTaskBodyPriority];

export const UpdateTaskBodyPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface UpdateTaskBody {
  title?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  projectId?: number | null;
  status?: UpdateTaskBodyStatus;
  priority?: UpdateTaskBodyPriority;
  /** @nullable */
  dueDate?: string | null;
}

export type UpdateTaskStatusBodyStatus =
  (typeof UpdateTaskStatusBodyStatus)[keyof typeof UpdateTaskStatusBodyStatus];

export const UpdateTaskStatusBodyStatus = {
  todo: "todo",
  in_progress: "in_progress",
  done: "done",
} as const;

export interface UpdateTaskStatusBody {
  status: UpdateTaskStatusBodyStatus;
}

export interface DashboardSummary {
  totalTasks: number;
  todoCount: number;
  inProgressCount: number;
  doneCount: number;
  totalProjects: number;
  highPriorityCount: number;
  dueSoonCount: number;
  completionRate: number;
}

export interface ActivityItem {
  id: number;
  taskId: number;
  taskTitle: string;
  action: string;
  /** @nullable */
  projectName?: string | null;
  /** @nullable */
  projectColor?: string | null;
  createdAt: string;
}

export type ListProjectTasksParams = {
  status?: ListProjectTasksStatus;
  priority?: ListProjectTasksPriority;
};

export type ListProjectTasksStatus =
  (typeof ListProjectTasksStatus)[keyof typeof ListProjectTasksStatus];

export const ListProjectTasksStatus = {
  todo: "todo",
  in_progress: "in_progress",
  done: "done",
} as const;

export type ListProjectTasksPriority =
  (typeof ListProjectTasksPriority)[keyof typeof ListProjectTasksPriority];

export const ListProjectTasksPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export type ListTasksParams = {
  status?: ListTasksStatus;
  priority?: ListTasksPriority;
  projectId?: number;
};

export type ListTasksStatus =
  (typeof ListTasksStatus)[keyof typeof ListTasksStatus];

export const ListTasksStatus = {
  todo: "todo",
  in_progress: "in_progress",
  done: "done",
} as const;

export type ListTasksPriority =
  (typeof ListTasksPriority)[keyof typeof ListTasksPriority];

export const ListTasksPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;