# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Task Manager SaaS API
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityItem {
  id: number;
  taskId: number;
  taskTitle: string;
  action: string;
  /** @nullable */
  projectName?: string | null;
  /** @nullable */
  projectColor?: string | null;
  createdAt: Date;
}