# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Task Manager SaaS API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalTasks: number;
  todoCount: number;
  inProgressCount: number;
  doneCount: number;
  totalProjects: number;
  highPriorityCount: number;
  dueSoonCount: number;
  completionRate: number;
}